requirements.txt
attrdict==2.0.1
base58==1.0.3
certifi==2019.6.16
chardet==3.0.4
cytoolz==0.10.0
Django==2.2.4
djangorestframework==3.10.2
eth-abi==2.0.0
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.6.4
hexbytes==0.2.0
idna==2.8
ipfshttpclient==0.4.12
jsonschema==2.6.0
lru-dict==1.1.6
multiaddr==0.0.8
netaddr==0.7.19
parsimonious==0.8.1
protobuf==3.9.1
pycryptodome==3.8.2
pytz==2019.2
requests==2.22.0
rlp==1.1.0
six==1.12.0
sqlparse==0.3.0
toolz==0.10.0
urllib3==1.25.3
varint==1.0.2
web3==5.0.1
websockets==7.0
